+++ rtl/files.f
@@ -4,4 +4,5 @@ rtl/rv64_decoder.sv
 rtl/rv64_imm_gen.sv
 rtl/rv64_regfile.sv
 rtl/rv64_mdu.sv
+rtl/rv64_amo.sv
 rtl/rv64_core.sv